// foods/src/foods.interface.ts
export enum EFoodMessage {
  FOOD_RECOVERED = 'Foods recovered',
  SECTION_CREATED = 'Section was properly created',
  EXTRA_CREATED = 'Extra was properly created',
  FOOD_CREATED = 'Food was properly created',
  FOOD_UPDATED = 'Food was properly updated',
  EXTRA_DELETED = 'Extra was properly deleted',
  SECTION_DELETED = 'Section was properly deleted',
  FOOD_DELETED = 'Food was properly deleted',
}